requirements.txt
gspread==5.7.2
numpy==1.24.3
oauth2client==4.1.3
pandas==2.0.2
Pillow==9.5.0
psycopg2_binary==2.9.6
streamlit==1.23.1
streamlit_authenticator==0.2.1
altair<=5.0.1
openpyxl==3.1.2
openpyxl>=3